[FLINK-17938] Move YarnTestBase.findFile and YarnTestBase.RootDirFilenameFilter to o.a.f.y.u.TestUtils
Moving findFile and RootDirFilenameFilter to o.a.f.y.u.TestUtils allows to run the UtilsTest suite without
requiring that yarn.classpath has been generated. The latter is required whenever one uses the YarnTestBase
class because it reads this file when the class is being loaded.
